+++ Cargo.toml
@@ -113,6 +113,9 @@ rspirv = { git = "https://github.com/tracel-ai/tracel-rspirv.git", rev = "9b4037
 ash = "0.38"
 tracel-ash = { git = "https://github.com/tracel-ai/tracel-ash.git", rev = "cedd894bbdb03f0635900e7c0d317020fdd38263" }
 
+# Build deps
+cfg_aliases = "0.2.1"
+
 [profile.dev]
 debug = 0     # Speed up compilation time and not necessary.
 opt-level = 0
